[weg2 flash-next WP3b] QSA DCP merge output cast back to the query dtype (fn1x: 'float != BFloat16' in o_proj on every rank)
fn1x (cyankiwi, DCP=1, pread loader): loads 104/109/167 s, ready, the QSA
rows path ran on all three ranks through the needle prefill and died after
2.3 min in the output projection: cp_lse_ag_out_ar_mha_uneven scales the
partials in fp32 and returns fp32. _attend_rows now returns q.dtype. Test
hermetic with stubbed gather/kernel/merge.

+++ python/sglang/srt/layers/attention/qwen_sparse_attn_backend.py
@@ -397,7 +397,10 @@ class QwenSparseAttnBackend(AttentionBackend):
         counts = self._dcp_group_q_head_counts(q.shape[1])
         q_all = cp_all_gather_heads_uneven(q.contiguous(), group, counts)
         out, lse = sparse_attn_rows_triton(q_all, k_pool, v_pool, rows, layer.scaling)
-        return cp_lse_ag_out_ar_mha_uneven(out, lse, group, counts)
+        # The merge scales in fp32 and hands back fp32 (fn1x 2026-09-16: every
+        # rank died in o_proj with 'float != BFloat16'); the layer's output
+        # projection expects the query dtype.
+        return cp_lse_ag_out_ar_mha_uneven(out, lse, group, counts).to(q.dtype)
 
     @staticmethod
     def _is_speculative_paged_mode(forward_mode) -> bool:
